requirements.txt
annotated-types==0.7.0
anyio==4.10.0
argcomplete==1.10.0
beautifulsoup4==4.8.0
chardet==3.0.4
click==8.2.1
docx2txt==0.8
EbookLib==0.17.1
extract-msg==0.23.1
fastapi==0.116.1
h11==0.16.0
idna==3.10
IMAPClient==2.1.0
lxml==6.0.0
olefile==0.46
pdfminer.six==20181108
pillow==11.3.0
pycryptodome==3.23.0
pydantic==2.11.7
pydantic_core==2.33.2
python-docx==1.2.0
python-multipart==0.0.20
python-pptx==0.6.18
pytz==2025.2
six==1.12.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
SpeechRecognition==3.8.1
starlette==0.47.2
textract==1.6.3
typing-inspection==0.4.1
typing_extensions==4.14.1
tzlocal==1.5.1
uvicorn==0.35.0
xlrd==1.2.0
xlsxwriter==3.2.5
